desktop: Use preload option for exposing specific things in main process
We only want to allow specific things from the main world to be
exposed to the render process, thus this patch adds a preload script
which allows exposing specific functionalities to the render process.
Also add locale as a valid channel. This patch also adds preload.js to
electron build.
This patch also disables node integration because enabling
nodeIntegration that allows outsiders to run any node script which can
be a security issue, we should use contextIsolation to provide what all
methods we want to expose from the main World.